// helpers/types.ts
export interface SymbolMap<T> {
  [symbol: string]: T;
}
export enum eTenderly {
  fork = "tenderly-fork",
}

export type eNetwork =
  | eEthereumNetwork
  | eTenderlyNetwork
  | eMantleNetwork
  | eTelosNetwork;

type eTenderlyNetwork = "tenderly";

export enum eEthereumNetwork {
  buidlerevm = "buidlerevm",
  kovan = "kovan",
  ropsten = "ropsten",
  main = "main",
  coverage = "coverage",
  hardhat = "hardhat",
  tenderly = "tenderly",
  rinkeby = "rinkeby",
  goerli = "goerli",
  sepolia = "sepolia",
}

export enum eMantleNetwork{
  mantleTestnet = "mantleTestnet",
  mantle = "mantle"
}

export enum eTelosNetwork{
  telosTestnet = "telosTestnet",
  telos = "telos"
}
// ...
export type tEthereumAddress = string;
// ...
export interface IReserveParams
  extends IReserveBorrowParams,
    IReserveCollateralParams {
  aTokenImpl: eContractid;
  reserveFactor: string;
  supplyCap: string;
  strategy: IInterestRateStrategyParams;
}

export interface IStrategy {
  [key: string]: IInterestRateStrategyParams;
}
// ...
export type iParamsPerNetwork<T> = {
  [k in eNetwork]?: T;
};
// ...
export interface EMode {
  id: string;
  ltv: string;
  liquidationThreshold: string;
  liquidationBonus: string;
  label: string;
  oracleId?: string;
  assets: string[];
}

export interface OracleConfig {
  kind: "PythOracle" | "DIAOracle" | "PriceOracle";
  address: string;
  config: IDiaFeedKey | IPythFeedID | ITokenAddress;
  OracleQuoteUnit: string;
  OracleQuoteCurrencyAddress: tEthereumAddress;
}

export interface IBaseConfiguration {
  SynthexAddress?: tEthereumAddress;
  MarketId: string;
  ATokenNamePrefix: string;
  StableDebtTokenNamePrefix: string;
  VariableDebtTokenNamePrefix: string;
  SymbolPrefix: string;
  ProviderId: number;
  PythId?: string;
  TestnetMarket?: boolean;
  ProviderRegistryOwner?: iParamsPerNetwork<tEthereumAddress | undefined>;

  FallbackOracle?: OracleConfig;
  PriceOracle: OracleConfig;

  WrappedTokenGateway?: iParamsPerNetwork<tEthereumAddress>;
  ReserveFactorTreasuryAddress: iParamsPerNetwork<tEthereumAddress>;
  StableDebtTokenImplementation?: iParamsPerNetwork<tEthereumAddress>;
  VariableDebtTokenImplementation?: iParamsPerNetwork<tEthereumAddress>;
  ReserveAssets?: iParamsPerNetwork<SymbolMap<tEthereumAddress>>;
  
  ReservesConfig: SymbolMap<IReserveParams>;
  WrappedNativeTokenSymbol: string; // ex: WETH or WMATIC
  IncentivesConfig: IncentivesConfig;
  EModes: SymbolMap<EMode>;
  L2PoolEnabled?: iParamsPerNetwork<boolean>;
  StkAaveProxy?: iParamsPerNetwork<tEthereumAddress>;
  ParaswapRegistry?: iParamsPerNetwork<tEthereumAddress>;
  FlashLoanPremiums: {
    total: number;
    protocol: number;
  };
  RateStrategies: IStrategy;
}

export interface ICommonConfiguration extends IBaseConfiguration {}
// ...
export interface IncentivesConfig {
  enabled: iParamsPerNetwork<boolean>;
  rewards: iParamsPerNetwork<SymbolMap<tEthereumAddress>>;
  rewardsOracle: iParamsPerNetwork<SymbolMap<tEthereumAddress>>;
  incentivesInput: iParamsPerNetwork<RewardsConfigInput[]>;
}